refactor: Add ability to specify plan file path in options (#35)
* refactor: Add ability to specify plan file path in options

Adds a new option function `WithPlanFile` that allows users to specify the
path to the plan file. The `planFile` field is added to the `Options` struct
to store this value. This enhancement provides more flexibility in managing
the plan file location.

* refactor: Add ability to specify plan file path in options

Adds a new option function `WithPlanFile` that allows users to specify the
path to the plan file. The `planFile` field is added to the `Options` struct
to store this value. This enhancement provides more flexibility in managing
the plan file location.